backend/app/requirements.txt
fastapi
uvicorn
python-multipart 
openai>=1.0.0
python-dotenv